+++ docs/src/pages/guide/reference/observability/_meta.js
@@ -1,8 +1,0 @@
-const meta = {
-  "create-logger": "createLogger()",
-  logger: "Logger",
-  "create-combinedlogger": "combineLoggers()",
-  "otel-config": "OtelConfig",
-};
-
-export default meta;