+++ docs/src/pages/guide/workflows/_meta.js
@@ -1,6 +1,0 @@
-const meta = {
-  "00-overview": "Overview",
-  "01-scraper": "Guide: Scraper",
-};
-
-export default meta;