+++ examples/dane/src/index.ts
@@ -3,8 +3,10 @@ import { Command } from 'commander';
 import dotenv from 'dotenv';
 
 import { commitMessageCommand } from './commands/commit-message.js';
+import { configCommand } from './commands/config.js';
 import { issueLabelerCommand } from './commands/issue-labeler.js';
 import { message } from './commands/message.js';
+import { publishPackages } from './commands/publish-packages.js';
 
 dotenv.config();
 
@@ -14,8 +16,18 @@ const program = new Command();
 
 program.command('chat').action(message);
 
-program.command('issue-labeler').action(issueLabelerCommand);
+program
+  .command('issue-labeler')
+  .description('Automatically label GitHub issues based on their content and context')
+  .action(issueLabelerCommand);
 
-program.command('commit').action(commitMessageCommand);
+program
+  .command('commit')
+  .description('Create a sensible commit message based on the changes made')
+  .action(commitMessageCommand);
+
+program.addCommand(configCommand);
+
+program.command('publish').description('Publish packages to the registry').action(publishPackages);
 
 program.parse(process.argv);