+++ packages/cli/src/index.ts
@@ -1,12 +1,9 @@
 #! /usr/bin/env node
-import * as p from '@clack/prompts';
 import { Command } from 'commander';
 import color from 'picocolors';
 
 import { PosthogAnalytics } from './analytics/index.js';
-import { createNewAgent } from './commands/agents/create-new-agent.js';
-import { listAgents } from './commands/agents/list-agent.js';
-import { updateAgentIndexFile } from './commands/agents/update-agent-file.js';
+import { create } from './commands/create/create.js';
 import { cloudflareDeploy, netlifyDeploy, vercelDeploy } from './commands/deploy/index.js';
 import { dev } from './commands/dev.js';
 import { add } from './commands/engine/add.js';
@@ -15,7 +12,7 @@ import { generate } from './commands/engine/generate.js';
 import { migrate } from './commands/engine/migrate.js';
 import { up } from './commands/engine/up.js';
 import { init } from './commands/init/init.js';
-import { checkAndInstallCoreDeps, checkPkgJsonAndCreateStarter, interactivePrompt } from './commands/init/utils.js';
+import { checkAndInstallCoreDeps, checkPkgJson, interactivePrompt } from './commands/init/utils.js';
 import { DepsService } from './services/service.deps.js';
 import { findApiKeys } from './utils/find-api-keys.js';
 import { getEnv } from './utils/get-env.js';
@@ -46,23 +43,56 @@ program
   });
 
 program
+  .command('create')
+  .description('Create a new Mastra project')
+  .option('--default', 'Quick start with defaults(src, OpenAI, no examples)')
+  .option('-c, --components <components>', 'Comma-separated list of components (agents, tools, workflows)')
+  .option('-l, --llm <model-provider>', 'Default model provider (openai, anthropic, or groq))')
+  .option('-e, --example', 'Include example code')
+  .action(async args => {
+    await analytics.trackCommandExecution({
+      command: 'create',
+      args,
+      execution: async () => {
+        if (args.default) {
+          await create({
+            components: ['agents', 'tools', 'workflows'],
+            llmProvider: 'openai',
+            addExample: false,
+          });
+        }
+        await create({
+          components: args.components,
+          llmProvider: args.llm,
+          addExample: args.example,
+        });
+      },
+    });
+  });
+
+program
   .command('init')
-  .description('Initialize a new Mastra project')
+  .description('Initialize Mastra in your project')
   .option('--default', 'Quick start with defaults(src, OpenAI, no examples)')
   .option('-d, --dir <directory>', 'Directory for Mastra files to (defaults to src/)')
   .option('-c, --components <components>', 'Comma-separated list of components (agents, tools, workflows)')
   .option('-l, --llm <model-provider>', 'Default model provider (openai, anthropic, or groq))')
   .option('-e, --example', 'Include example code')
-  .option('-ne, --no-example', 'Skip example code')
   .action(async args => {
     await analytics.trackCommandExecution({
       command: 'init',
       args,
       execution: async () => {
-        await checkPkgJsonAndCreateStarter();
+        await checkPkgJson();
         await checkAndInstallCoreDeps();
 
-        if (!Object.keys(args).length) return interactivePrompt();
+        if (!Object.keys(args).length) {
+          const result = await interactivePrompt();
+          await init({
+            ...result,
+          });
+          return;
+        }
 
         if (args?.default) {
           init({
@@ -70,7 +100,6 @@ program
             components: ['agents', 'tools', 'workflows'],
             llmProvider: 'openai',
             addExample: false,
-            showSpinner: true,
           });
           return;
         }
@@ -81,7 +110,6 @@ program
           components: componentsArr,
           llmProvider: args.llm,
           addExample: args.example,
-          showSpinner: true,
         });
         return;
       },
@@ -174,49 +202,6 @@ engine
             `Run ${color.blueBright('mastra engine add')} to get started with a Postgres DB in a docker container`,
           );
         }
-      },
-    });
-  });
-
-const agent = program.command('agent').description('Manage Mastra agents');
-
-agent
-  .command('new')
-  .description('Create a new agent')
-  .option('-d, --dir <dir>', 'Path to your mastra folder')
-  .action(async args => {
-    await analytics.trackCommandExecution({
-      command: 'agent new',
-      args: {},
-      execution: async () => {
-        const result = await createNewAgent({ dir: args?.dir });
-        if (!result) return;
-        await updateAgentIndexFile({ newAgentName: result, dir: args?.dir });
-      },
-    });
-  });
-
-agent
-  .command('list')
-  .description('List all agents')
-  .option('-d, --dir <dir>', 'Path to your mastra folder')
-  .action(async args => {
-    await analytics.trackCommandExecution({
-      command: 'agent list',
-      args,
-      execution: async () => {
-        const agents = await listAgents({ dir: args?.dir });
-
-        if (!agents.length) {
-          p.note('No Agents...');
-          return;
-        }
-
-        p.intro(color.bgCyan(color.black(' Agents List ')));
-
-        agents.forEach((agent, index) => {
-          logger.log(`${index + 1}. ${color.blue(agent)}`);
-        });
       },
     });
   });