+++ docs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # docs
 
+## 0.1.1-alpha.5
+
+### Patch Changes
+
+- 8ae2bbc: Dane publishing
+- 0bd142c: Fixes learned from docs
+- ee4de15: Dane fixes
+
+## 0.1.1-alpha.4
+
+### Patch Changes
+
+- 0091799: Add dev and deploy mastra commands to CLI references in documentation, update build successful message in dev command
+
 ## 0.1.1-alpha.3
 
 ### Patch Changes