+++ examples/vnext-playground/CHANGELOG.md
@@ -1,5 +1,31 @@
 # examples-vnext-playground
 
+## 1.0.1-alpha.24
+
+### Patch Changes
+
+- Updated dependencies [b726bf5]
+  - @mastra/core@0.1.27-alpha.48
+  - @mastra/engine@0.0.5-alpha.42
+
+## 1.0.1-alpha.23
+
+### Patch Changes
+
+- Updated dependencies [f6ba259]
+  - @mastra/core@0.1.27-alpha.47
+  - @mastra/engine@0.0.5-alpha.41
+
+## 1.0.1-alpha.22
+
+### Patch Changes
+
+- Updated dependencies [8ae2bbc]
+- Updated dependencies [0bd142c]
+- Updated dependencies [ee4de15]
+  - @mastra/engine@0.0.5-alpha.40
+  - @mastra/core@0.1.27-alpha.46
+
 ## 1.0.1-alpha.21
 
 ### Patch Changes