+++ integrations/firecrawl/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @mastra/firecrawl
 
+## 1.0.4-alpha.34
+
+### Patch Changes
+
+- Updated dependencies [b726bf5]
+  - @mastra/core@0.1.27-alpha.48
+  - @mastra/rag@0.0.2-alpha.32
+
+## 1.0.4-alpha.33
+
+### Patch Changes
+
+- Updated dependencies [f6ba259]
+  - @mastra/core@0.1.27-alpha.47
+  - @mastra/rag@0.0.2-alpha.31
+
+## 1.0.4-alpha.32
+
+### Patch Changes
+
+- 8ae2bbc: Dane publishing
+- 0bd142c: Fixes learned from docs
+- ee4de15: Dane fixes
+- Updated dependencies [8ae2bbc]
+- Updated dependencies [0bd142c]
+- Updated dependencies [ee4de15]
+  - @mastra/core@0.1.27-alpha.46
+  - @mastra/rag@0.0.2-alpha.30
+
+## 1.0.4-alpha.31
+
+### Patch Changes
+
+- 9897f48: Use conventional commits
+
 ## 1.0.4-alpha.30
 
 ### Patch Changes