+++ packages/memory/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @mastra/memory
 
+## 0.0.2-alpha.28
+
+### Patch Changes
+
+- Updated dependencies [b726bf5]
+  - @mastra/core@0.1.27-alpha.48
+
+## 0.0.2-alpha.27
+
+### Patch Changes
+
+- Updated dependencies [f6ba259]
+  - @mastra/core@0.1.27-alpha.47
+
+## 0.0.2-alpha.26
+
+### Patch Changes
+
+- 8ae2bbc: Dane publishing
+- 0bd142c: Fixes learned from docs
+- ee4de15: Dane fixes
+- Updated dependencies [8ae2bbc]
+- Updated dependencies [0bd142c]
+- Updated dependencies [ee4de15]
+  - @mastra/core@0.1.27-alpha.46
+
+## 0.0.2-alpha.25
+
+### Patch Changes
+
+- 3220d26: Fix lastStep error in agent stream
+
 ## 0.0.2-alpha.24
 
 ### Patch Changes